Player/requirements.txt
cryptography~=42.0.5
pywebview~=5.4